Fix type definitions

+++ proc.h
@@ -56,7 +56,7 @@ typedef enum ReleaseReason ReleaseReason;
 void k_initProcesses(ProcInfo *procInfo);
 uint32_t k_releaseProcessor(ProcInfo *procInfo, ReleaseReason reason);
 uint32_t k_setProcessPriority(ProcInfo *procInfo, ProcId pid, uint8_t priority);
-uint32_t k_getProcessPriority(ProcInfo *procInfo, ProcId pid);
+int16_t k_getProcessPriority(ProcInfo *procInfo, ProcId pid);
 
 #endif
